+++ CHANGELOG.md
@@ -12,3 +12,23 @@ the following form: YYYY.0M.0D.
 - Designed a CrossyRoadGame component
 - Designed a Playlist component
 - Designed a EntertainmentTracker component
+
+## 2025.10.09
+
+### Added
+
+- Designed a proof of concept for crossy road game component
+
+### Updated
+
+- Changed design to include 5x5 grid, player movement, obstacle movement, collision checking, and the grid
+
+## 2025.10.23
+
+### Added
+
+- Designed kernel and enhanced interfaces for crossy road game component
+
+### Updated
+
+- Changed design to include documented interface contracts